+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.2.0
+
+- Added option to select choices with modifier keys (instead of
+  the usual `enter` or `space` keys). 
+
+## 0.1.0
+
+- First public version, with improved API.
+
 ## 0.0.1
 
-- Initial version, created by Stagehand
+- Initial version.